Add RUN_COVERAGE flag.
This will only generate a test coverage report when 'leeroy' is not
passed to the build script.

Fixes bug 1044072

+++ Makefile
@@ -6,7 +6,7 @@ PREFIX=/data/socorro
 ABS_PREFIX = $(shell readlink -f $(PREFIX))
 VIRTUALENV=$(CURDIR)/socorro-virtualenv
 PYTHONPATH = "."
-NOSE = $(VIRTUALENV)/bin/nosetests socorro -s --with-xunit
+NOSE = $(VIRTUALENV)/bin/nosetests socorro -s
 SETUPDB = $(VIRTUALENV)/bin/python ./socorro/external/postgresql/setupdb_app.py
 COVERAGE = $(VIRTUALENV)/bin/coverage
 PYLINT = $(VIRTUALENV)/bin/pylint
@@ -33,10 +33,14 @@ test: bootstrap
 	PYTHONPATH=$(PYTHONPATH) $(SETUPDB) --database_name=socorro_migration_test --database_username=$(database_username) --database_hostname=$(database_hostname) --database_password=$(database_password) --database_port=$(DB_PORT) --database_superusername=$(database_superusername) --database_superuserpassword=$(database_superuserpassword) --dropdb --logging.stderr_error_logging_level=40 --unlogged
 	PYTHONPATH=$(PYTHONPATH) $(VIRTUALENV)/bin/alembic -c config/alembic.ini downgrade -1
 	PYTHONPATH=$(PYTHONPATH) $(VIRTUALENV)/bin/alembic -c config/alembic.ini upgrade +1
-	# run tests with coverage
-	rm -f coverage.xml
-	$(ENV) $(PG_RESOURCES) $(RMQ_RESOURCES) $(ES_RESOURCES) PYTHONPATH=$(PYTHONPATH) $(COVERAGE) run $(NOSE)
-	$(COVERAGE) xml
+	if [ -z $(RUN_COVERAGE) ]; then 
+	  # run tests with coverage
+	  rm -f coverage.xml
+	  $(ENV) $(PG_RESOURCES) $(RMQ_RESOURCES) $(ES_RESOURCES) PYTHONPATH=$(PYTHONPATH) $(COVERAGE) run $(NOSE) --with-xunit
+	  $(COVERAGE) xml
+	else
+	  $(ENV) $(PG_RESOURCES) $(RMQ_RESOURCES) $(ES_RESOURCES) PYTHONPATH=$(PYTHONPATH) run $(NOSE)
+	fi
 	# test webapp
 	cd webapp-django; ./bin/jenkins.sh
